# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Clinica medica",
        "description": "",
        "version": "1.0.0-beta"
    },
    "paths": {
        "/api/appointments/{appointment}": {
            "get": {
                "summary": "Get appointment by ID",
                "description": "Returns a single appointment",
                "operationId": "getAppointmentById",
                "parameters": [
                    {
                        "name": "appointment",
                        "in": "path",
                        "description": "ID of the appointment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Appointment located successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/patienties/{patient}": {
            "get": {
                "summary": "Get patient by ID",
                "description": "Returns a single patient",
                "operationId": "getPatientById",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "description": "ID of the patient",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Patient located successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Patient"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Patient not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Appointment": {
                "title": "Appointment",
                "properties": {
                    "name": {
                        "description": "Name of the appointment",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Patient": {
                "title": "Patient"
            }
        }
    }
}